chore(core): move search input v2
